feat(web): release the day-granular list and delete VITE_NAV_V2
