+++ frontend/.env.example
@@ -12,8 +12,3 @@ VITE_API_URL=
 # Enable fetching and rendering of events from registered publisher feeds.
 # When unset or "false", only chq.org primary events are loaded.
 VITE_ENABLE_PUBLISHER_FEEDS=false
-
-# Date-navigation phase 2: day-granular list growth, automatic forward
-# expansion past the end of the current scope, and a "Show earlier" control.
-# When unset or "false", the list behaves exactly as it did before.
-VITE_NAV_V2=false
