+++ frontend/src/app/about/aboutContent.ts
@@ -352,8 +352,8 @@ export const WEB_FEATURES: Feature[] = [
     blurb: 'Three buttons cover the questions people actually ask.' },
   { id: 'web-adaptive', group: 'Dates & weeks', title: 'The “next” window grows to fit', notObvious: true,
     blurb: 'Instead of a fixed number of days, it widens until it has a useful number of events to show.' },
-  { id: 'web-show-next-day', group: 'Dates & weeks', title: 'Show next day', notObvious: true,
-    blurb: 'One click extends the window a day at a time, as far as you want to look. The button names the date it is about to add.' },
+  { id: 'web-show-more', group: 'Dates & weeks', title: 'The list keeps going', notObvious: true,
+    blurb: 'Scroll past the end of what you asked for and the next day loads on its own. Going backwards stays deliberate — a “Show earlier” button at the top names the day it is about to add.' },
   { id: 'web-week-strip', group: 'Dates & weeks', title: 'Pick weeks from the strip',
     blurb: 'Every week of the season, selectable directly.' },
   { id: 'web-week-drag', group: 'Dates & weeks', title: 'Drag to select a range of weeks', notObvious: true,
